Update virtualbox and vagrant installer
Validate sudo before starting to download deb package so one does not
need to supervise the whole installation (fire and forget).
Replace wget with curl so the amount of necessary applications is low.
Move the `then` clauses to the next line for better readability.
Add the permission notice to the help text before the run.